+++ austrakka/components/metadata/__init__.py
@@ -2,7 +2,7 @@
 from io import BufferedReader
 import click
 
-from austrakka.utils.options import species as species_opt
+from austrakka.utils.options import opt_species
 from .funcs import add_metadata
 
 
@@ -15,7 +15,7 @@ def metadata(ctx):
 
 @metadata.command('add')
 @click.argument('file', type=click.File('rb'))
-@species_opt
+@opt_species
 def submission_add(file: BufferedReader, species: int):
     """Upload metadata submission to AusTrakka"""
     add_metadata(file, species)