+++ austrakka/components/sequence/funcs.py
@@ -1,3 +1,4 @@
+import os
 from io import BufferedReader
 from typing import Tuple
 from os import path
@@ -12,14 +13,22 @@ from loguru import logger
 from austrakka.utils.exceptions import FailedResponseException
 from austrakka.utils.misc import logger_wraps
 from austrakka.utils.api import call_api
+from austrakka.utils.api import call_api_raw
+from austrakka.utils.api import call_get_api
 from austrakka.utils.api import post
 from austrakka.utils.api import RESPONSE_TYPE_ERROR
 from austrakka.utils.paths import SEQUENCE_PATH
+from austrakka.utils.paths import SAMPLE_BY_SPECIES_PATH
+from austrakka.utils.paths import SAMPLE_DOWNLOAD_INFO_PATH
 from austrakka.utils.output import create_response_object
 from austrakka.utils.output import log_response
+from austrakka.utils.fs import create_dir
 
-FASTA_UPLOAD = 'Fasta'
-FASTQ_UPLOAD = 'Fastq'
+FASTA = 'Fasta'
+FASTQ = 'Fastq'
+DOWNLOAD = 'download'
+
+ALL_READS = "-1"
 
 FASTQ_CSV_SAMPLE_ID = 'sampleId'
 FASTQ_CSV_FILENAME_1 = 'filename1'
@@ -32,7 +41,7 @@ FASTQ_CSV_SPECIES = 'species'
 def add_fasta_submission(files: Tuple[BufferedReader], csv: BufferedReader):
     call_api(
         method=post,
-        path=path.join(SEQUENCE_PATH, FASTA_UPLOAD),
+        path=path.join(SEQUENCE_PATH, FASTA),
         body=[('files[]', (file.name, file)) for file in files]
         + [('files[]', (csv.name, csv))],
         multipart=True,
@@ -83,7 +92,7 @@ def add_fastq_submission(files: Tuple[BufferedReader], csv: BufferedReader):
         try:
             call_api(
                 method=post,
-                path=path.join(SEQUENCE_PATH, FASTQ_UPLOAD),
+                path=path.join(SEQUENCE_PATH, FASTQ),
                 body=sample_files,
                 multipart=True,
                 custom_headers=custom_headers,
@@ -93,6 +102,20 @@ def add_fastq_submission(files: Tuple[BufferedReader], csv: BufferedReader):
             log_response(ex.parsed_resp)
         except Exception as ex:
             raise ex from ex
+
+
+@logger_wraps()
+def download_fastq(species: str, output_dir: str, read: str):
+
+    # fetch sample list
+    if not os.path.exists(output_dir):
+        create_dir(output_dir)
+
+    data = fetch_samples_names_by_species(species)
+    samples_names = list(data)
+    throw_if_empty(samples_names, f'No samples found for species: {species}')
+    samples_seq_info = fetch_seq_download_info(samples_names)
+    download_fastq_for_each_sample(output_dir, samples_seq_info, read)
 
 
 def _validate_fastq_submission(
@@ -160,3 +183,78 @@ def _validate_fastq_submission(
             ))
 
     return messages
+
+
+def download_fastq_for_each_sample(
+    output_dir: str,
+    samples_seq_info: list[tuple[any, any]],
+    read: str
+):
+    for ssi in samples_seq_info:
+        sample_name = ssi[0]
+
+        for seq_dto in ssi[1]:
+            dto_read = str(seq_dto['read'])
+
+            # When read is -1, it means take both.
+            if read not in (dto_read, ALL_READS):
+                continue
+
+            filename = seq_dto['fileName']
+            sample_dir = os.path.join(output_dir, sample_name)
+            file_path = os.path.join(sample_dir, filename)
+
+            if os.path.exists(file_path):
+                logger.warning(
+                    f'Found a local copy of {filename}.  Skipping...')
+                continue
+
+            try:
+                resp = call_api_raw(
+                    path=path.join(
+                        SEQUENCE_PATH,
+                        DOWNLOAD,
+                        FASTQ,
+                        sample_name,
+                        dto_read,
+                    ),
+                    stream=True,
+                )
+
+                if not os.path.exists(sample_dir):
+                    create_dir(sample_dir)
+
+                with open(file_path, 'wb') as file:
+                    for chunk in resp.iter_content(chunk_size=128):
+                        file.write(chunk)
+
+                logger.success(f'Downloaded: {filename} To: {file_path}')
+
+            except FailedResponseException as ex:
+                logger.error(
+                    f'Error while downloading file for sample: '
+                    f'{sample_name}, read: {dto_read}. \n{ex.message}')
+
+
+def fetch_seq_download_info(sample_names):
+    samples_seq_info = []
+    for name in sample_names:
+        info = call_get_api(
+            path=path.join(SAMPLE_DOWNLOAD_INFO_PATH, name),
+        )
+        samples_seq_info.append((name, info))
+    return samples_seq_info
+
+
+def throw_if_empty(a_list: list, msg: str):
+    if not a_list:
+        raise FailedResponseException(
+            create_response_object(
+                msg, RESPONSE_TYPE_ERROR))
+
+
+def fetch_samples_names_by_species(species):
+    samples = call_get_api(
+        path=path.join(SAMPLE_BY_SPECIES_PATH, species),
+    )
+    return samples