+++ austrakka/utils/enums/seq.py
@@ -1,0 +1,3 @@
+
+FASTQ_UPLOAD_TYPE = 'fastq'
+FASTA_UPLOAD_TYPE = 'fasta'