+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [0.6.1] - 2022-05-04
+### Changed
+- Changed to require --outdir option when using seq command.
+- Added examples to the seq command help section.
+-
+## [0.6.0] - 2022-05-03
+### Added
+- Support for downloading fastq files by species ID.
+
 ## [0.5.1] - 2022-05-04
 ### Changes
 - CI/CD upgrades.
@@ -106,4 +115,3 @@ and this project adheres to [Semantic Versioning].
 [0.4.0]: https://github.com/AusTrakka/austrakka2-cli/releases/tag/0.4.0
 [0.4.1]: https://github.com/AusTrakka/austrakka2-cli/releases/tag/0.4.1
 [0.5.0]: https://github.com/AusTrakka/austrakka2-cli/releases/tag/0.5.0
-[0.5.1]: https://github.com/AusTrakka/austrakka2-cli/releases/tag/0.5.1